// generated/api/models/transactionType.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * WolfBankGateway
 * OpenAPI spec version: 1.0
 */

export type TransactionType = (typeof TransactionType)[keyof typeof TransactionType];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const TransactionType = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4
} as const;